requirements.txt
openai
ib_insync
pandas
llama-index
jinja2
streamlit
python-dotenv
request